[clang-tidy] fix `bugprone-too-small-loop-variable` warnings

+++ src/cli/cli.cpp
@@ -5256,7 +5256,7 @@ void Interpreter::HandleMeshDiagDiscoverDone(otError aError, otMeshDiagRouterInf
         {
             OutputFormat(kIndentSize, "%u-links:{ ", linkQuality);
 
-            for (uint8_t id = 0; id < OT_ARRAY_LENGTH(aRouterInfo->mLinkQualities); id++)
+            for (uint8_t id = 0; id < static_cast<uint8_t>(OT_ARRAY_LENGTH(aRouterInfo->mLinkQualities)); id++)
             {
                 if (aRouterInfo->mLinkQualities[id] == linkQuality)
                 {
